// app/types/Student.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
export type Student = {
  id?: string;
  firstName: string;
  lastName: string;
  email: string;
  birthDate: any;
  createdAt: any;
  createdBy: string;
  city: string;
  country: string;
  university: string;
  major: string;
  program: string;
};